refact: KEYP-104: create the PoCommand class (replace PoBuilderParser)
